Decompile duplicates of CreateEntityFromEntity (#120)
The following functions, which are duplicates of CreateEntityFromEntity
in other zones, have been decompiled:

* func_80193538
* func_8018C854

Also includes a decompilation of the EntitySlograSpearProjectile
function, which was already done by Sonicdreamcaster, I believe.

This is my first attempt at a pull request, so I started off with some
easy ones to get familiar with the process. If there are any guidelines
I neglected to follow, let me know, and I'll keep them in mind for
future pull requests.