Split remaining DRA jump tables (#132)
This should split out the remaining DRA jump tables. This was done with
a script but I looked over most of the files and it seems correct to me.
The purpose is making decompiling jump table functions a little quicker.

I'm thinking about writing a script to split all the C files
accordingly, do we agree with that? There's about 95 remaining jump
table functions in DRA, so there would be about that many file splits.
This would save time when working on jump table functions since all the
splitting work would already be done.