WRP split some data and add new split tools (#109)
The objective is to separate the data that is usually written into `asm/*/data` into separate files. Those will initially be binary files.

There is also an utility that allows to decompress compressed data. The decompressed data is currently not read as we are missing a compression algorithm that matches the original one. I tried writing one in the past: it worked, it re-created similar files but sadly not the same ones. The decompression algorithm is there to allow to do some data mining and understand what the data is and how it is used.