Add scratch scraping to function_finder (#173)
This adds the ability to scrape decomp.me for functions. We just look
for ps1 and the same function name so collisions are possible. The
in-source entries are considered definitive and scraped entries are
labeled in the output so that it's known they may be less reliable.
There's many scratches that we don't have labeled in the source so this
should help avoid duplicated effort.

Here's an example of the output:

https://gist.github.com/sozud/2fa067f9434a326020f09f8fe6a2e75d