EntityEnemyBlood co-authored-by @nim-ka, @Xeeynamo, @pixel-stuck (#175)
This function is weird in various ways, for starters, i needed to remove
the void type from the prototype argument of MoveEntity(), because the
function needed 'self' to be passed to it for matching, that could have
been a mistake from the dev, since DestroyEntity() takes an Entity* for
example, so it's very likely someone made that mistake, and like we have
been encountering a lot of times, it seems like the devs weren't care
about doing prototypes for functions so the compiler would assume
MoveEntity(s32), making it possible to receive an argument without
erroring out. There's no other way that i or @nim-ka found to match the
function. The other possibility is that the function actually takes an
Entity* but it's unused, but i think this is unlikely.

The other thing is that the function needs to use 'x2' for the last
member used below to match, but in decomp.me compiler it needs to be
'y2' for the offset to be correct in the asm, i couldn't figure out why.

I'm leaving the already cleaned up scratch in case someone wants to mess
with it and try to figure out a solution for the MoveEntity() argument
needed, and once we're all sure this is what we want, i'll make the
other duplicates in another PR.

https://decomp.me/scratch/sPUOP