Match func_801B12E8 aka EntityPurpleBrickScrollingBackground (#196)
This pull request matches NZ0 `func_801B12E8`. This particular function
is responsible for handling one of the backdrops of the Alchemy
Laboratory that scrolls with the camera. It appears as a repeating
purplish-red brick pattern behind the foreground.

It is comprised of 15 textured primitives laid out in a 5-tall by 3-wide
grid.

- NZ0 - `func_801B12E8` - Renamed to
`EntityPurpleBrickScrollingBackground`

---------

Co-authored-by: Alejandro Asenjo Nitti <96613413+sonicdcer@users.noreply.github.com>