Queue CI actions by branch
This avoids the problem of merging multiple pull requests in a short
span. Usually what happens is the generate-duplicate-report is slow
and since it updates a branch in the repo there are occasiosn where
multiple actions would update the same branch concurrently. This will
always result in an execution failure. By disabling the concurrency it
is possible to fix this behaviour. Actions are queued by branch, job
name and event name refs-head-master-build-and-test-pull_request_target
which will still allow multiple PR jobs to run by their own and isolated