DRA func_80118810 and RIC func_8015F8F8, func_8015F96C
This function is the same as GetFreeDraEntity, but running in reverse
order.

This function has a duplicate in Ric, and therefore I have also done the
replacement of the duplicate.

While replacing the duplicate in Ric, I noticed that Ric also has a
duplicate of GetFreeDraEntity, but with less concise code. I therefore
also replaced GetFreeDraEntity into its duplicate in Ric.