DRA func_8012CB4C decompiled (#241)
Short and simple function, but took a bit to get rid of the goto's :)

This seems to be the input handler that handles moving the player left
and right in response to the left and right controller inputs. We also
add a new variable, D_8013842C. Its purpose is unclear; I will next
attempt to decompile some of the other functions using that variable.