Extract Alchemy Lab, decompile func_060e1ff8 (#240)
This adds the Alchemy Lab overlay and decompiles func_060e1ff8 to check
that it's working. func_060e1ff8 seems similar to
https://github.com/Xeeynamo/sotn-decomp/blob/31d04a46d3f9c258722cedc56627d0dd627cbf71/src/st/nz0/36DE4.c#L101
but it's not clear if it's the same function yet.

I think it probably makes sense to add one overlay of each type. Right
now there's no stage overlay extracted. I picked Alchemy Lab since it's
the furthest along on PSX.