DRA func_80117D3C matching (#243)
Another function working! This one was 99% done by Sozud, I just
recognized a problem in the conditionals that was causing it to use an
extra register.

Seems to be something related to the mist-form, since it checks for
pressing the controller's L1 button. Will try to decompile some of the
functions called by this one.

Co-authored  by: @sozud