DrawMenuStr now decompiled (#250)
This is a function which writes strings on the menu screens. Xeeynamo
had previously started a scratch for this, but I got it to be matching.
Code is ugly, but functional. Submitting this now as an operational
function; myself or others may make the logic more straightforward in
future.

No real changes outside the function itself; no new externs needed. Had
to change the function signature to use u8 instead of char in order to
get it to match.

As always, happy to make any requested changes in formatting or similar.

Thanks a bunch!