Switch ASPATCH to MASPSX (#254)
This PR replaces the ASPATCH command with my custom script named MASPSX.
It's very new, and so is likely to have bugs and/or missing
functionality. That said it's producing an :ok: build across all
overlays, and is matching a few functions that required psyq3.5/aspsx.

I have not attempted to match any of the functions that require rodata,
nor any of the functions that refused to compile / had diffs.

Note that I have added `-gcoff` to the `CC_FLAGS`, this will give you
line numbers in asm-differ.

----------------------

Thank you in advance for contributing to sotn-decomp.
Before submitting this PR, please make sure:

- [x] The PR is small and focuses to address a single task
- [x] `make all` reports all `OK`
- [x] `make format` reports no files changed
- [x] Have Actions enabled in your fork to run the auto-formatter