Changes to MenuContext (#255)
I am currently working on decompiling func_800F8F28. This relates to the
system pause menus, and makes heavy use of the MenuContext struct. This
calls several members of this struct, in ways that have revealed more
details about this struct, including some of its members which were
unknown in size or purpose previously.

This PR serves to adjust the struct, and also adjusts the few places in
existing code which accessed the unk4 and unk6 members (now known as
cursorW and cursorH).

This PR, on its own, serves minimal functional benefit. However, in the
interest of "One PR for one change", I wanted to submit this first,
rather than having my future PR for func_800F8F28 serve the purpose of
both decompiling the function, and changing the struct. The struct is
ready for changes now so I am submitting this first so it can be
considered separately.

Full list of changes to the struct:
unk4 and unk6 are now cursorW and cursorH.
Fixed typo on the list of addresses that had 0x10 on two separate struct
members
Split unk14 into unk14 and unk16; my func_800F8F28 accesses these two
locations individually.
Split unk1C into unk1C and unk1D; same deal there.

- [x] The PR is small and focuses to address a single task
- [x] `make all` reports all `OK`
- [x] `make format` reports no files changed
- [x] Have Actions enabled in your fork to run the auto-formatter