NP3 Some functions from Gurkha and Hammer (#268)
I'm still trying to decompile Gurkha but run into some problems with
some of his loops. I decided to take a break from it for a while so i
figured i should submit what i have for now so it doesn't get lost in
the 7th dimension.
Some of them will be refactored when the two main entities get
completely decompiled.