Change references to Equipment.damageScale to itemCategory (#275)
The entry damageScale directly corresponds to the in-game item category
list, which lists categories such as Sword, Food, and Shield. I noticed
this correspondence by looking at the equipment.json. This PR renames
all references to damageScale to now refer to itemCategory.

I also took the liberty of creating an enum to reference each category.
This copies the category names out of the in-game menu.