Decompile CEN func_801989AC and duplicates (#281)
This function appears to deal with a bunch of collision logic. I won't
pretend to understand it, but maybe someone after me can identify its
purpose and make any naming tweaks.

This function appears in 9 different overlays; code is identical in all
of them.

The function calls MoveEntity, which is also heavily duplicated.
However, for some reason MoveEntity within rwrp was not decompiled -
I've taken the liberty of copying the decompilation of all the other
MoveEntity instances into rwrp, and adding it to
config/symbols.us.strwrp.txt, and changing the other calls to use the
new name MoveEntity. Therefore, this PR does not strictly only address a
single task, but given how common MoveEntity is in the other overlays,
updating rwrp to match the others seems like a relatively innocent
change.

Big thank you to @sonicdcer for help resolving my final mismatch :)