Decompile RIC func_8015FB84; improve SubweaponDef struct (#283)
Primary purpose of this is to decompile the function. Testing in
debugger indicates that this runs every frame, and only has the
useHearts parameter set to 1 when the player actually uses their
subweapon. I applied similar emulator-driven logic to determine the
purpose of the second parameter which I identified as isItemCrash.

Through the decompiling of this function, I have also identified two
previously-unknown members of the SubweaponDef struct. I'm quite
confident about the heartCost, as it is the amount that is subtracted
from the player's hearts. I'm also fairly confident about the crashId,
since its logic corresponds directly to the isItemCrash parameter that
the function takes, and is the only reason I can think of that a
subweapon would point to another subweapon (note the double-reference to
D_80154688).

Updating the struct broke two other functions which referred to the
placeholder names for those struct members; I have therefore updated
those functions to use my new names (heartCost and crashId).