Decompile DRA func_8011A9D8 (#294)
Yet another random function decompiled.

This uses two flags in entity->flags. One was already listed in game.h,
the other was not. They are now each included. I also added a comment to
one of them mentioning its purpose; this could lead to a rename in the
future when more uses of it are found.