Decompile RIC func_8015D250 (#292)
Decompiled this function. There was a comment in 20920.C on this one's
IMPORT_ASM saying "// decompiling func_8015FB84 first is recommended to
discover struct" - my previous PR did that, so now this one was ready to
go!

The jump table here caused some issues which resulted in needing to
split 20920.c into two files. This now does that.

This is my first time changing file structures and changing splats to
have different C file output, so let me know if anything looks wrong :)

Thanks to @sozud and @sonicdcer for the help working out splitting.