InitRoomEntities and duplicates matching
This was a function I started decompiling months ago. Today I was
finally able to match it. Thanks to @sonicdcer for dealing the last reg
swap. I was stumbling too hard on it without success.

I had to change the type of LayoutObject* into u16* or it would not
match otherwise.

Co-authored-by: sonicdcer <sonicdcer@users.noreply.github.com>