Decompile RIC func_8015D1D0 (#295)
This one's logic is a little odd. It is only called by func_8015D250, in
one place. I'll be interested to learn how it's supposed to work and
what the unkB0 and unk6 entries are from those structs. In any case,
it's a strange but short function and therefore the code is
self-contained, no changes outside the function itself.

Note: With this PR, src/ric/20920.c is fully decompiled and contains no
INCLUDE_ASM macros!