Add richter.prg symbols script (#310)
@sonicdcer pointed out previously that RICHTER.PRG has what looks to be
linker symbol information embedded in it. I've written a script to sort
and dump it. Here's the output:
https://gist.github.com/sozud/4d6299329f2ef34f880e7ffdfe57f5a8

I've checked a few symbols previously and it seems to be the same build
or pretty close to it. I imagine a decent number of these symbols would
also apply to the PSX build but will be at different addresses.

Co-authored-by: sozud <sozud@users.noreply.github.com>