Split more of 0.BIN, apply symbols (#312)
This PR splits more of 0.BIN and applies the RICHTER.PRG symbols. I had
to fix a bug in saturn-splitter so that is updated. I haven't checked
the new sotn-specific symbols but all the sega library stuff matched.

Regarding adding the symbols to the symbol files in this and the
previous PR, right now, for stuff that isn't extern, it doesn't do
anything. I think mkst said previously we shouldn't add symbols that the
splitter knows about since it makes shiftability harder? My splitter
operates a bit differently from splat at this time but I think that part
would end up being about the same once it becomes more mature. Anyway I
don't think it needs to be done right now and it would be easy to
automate if we figure out it needs to change in the future.