Decompile DRA EntitySubwpnCrashCrossParticles (#322)
Fairly straightforward entity decompilation. This certainly seems to use
unk7C to store a time-to-live. It also doesn't appear to use the entity
extension for anything else, so maybe this should become a new
extension?

One thing I notice is that this is for "rising blue particles", but I
don't see a decrementing Y value that would indicate rising, so there
must be somewhere else that controls the behavior of these particles.