DRA func_8011A4D0 matching (#323)
This is the main large function that runs the updater for every entity.

Some of the code doesn't look great (not sure why we need 3 i values)
but it matches and that's what counts.

Please double-check my new externs in dra.h, I'm not sure if they're in
the right places. aAtariNuki is at 0x800E1AA4 and the others have their
addresses in their names.