CollectSubweapon + duplicates (#326)
This function was already decompiled for ST0, it was reviewed and merged
in a previous PR, The one for the MAD overlay is a duplicate of the same
func. As for the rest of the overlays, the function calls SetStep
instead of just modifying the Current Entity step and step_s directly,
so with that little change the "normal" version of the function is now
decompiled.