DRA UpdateAnim (#325)
This function previously was marked as a void, but code indicates that
it should be returning a value, so it does.

Thanks to @Xeeynamo for pointing out the similarity to func_80173C64 in
tt_000 - Maybe that one should be renamed to match this one?

Anyway, matching now due to a mix of my own code with code pulled from
func_80173C64. I managed to remove the extra variables in that function
so this seems like a minimal working setup.

The final block is very ugly and requires changing things to types that
don't seem to make sense, but at least it matches.