DRA func_800F4994 matching (#329)
This is a really cool one, there was a scratch of this at
https://decomp.me/scratch/bXnUC but that was before the PlayerStatus
struct was established. Now we have a version of this that is pretty
readable, which is great!

I created an enum for the stats - there was an unused struct for stats,
but given how often stats are used by number, I think an enum makes more
sense, so I swapped that out and it seems to work here in the section
with the relic checks.

I think I would like to make an enum of relic names to continue making
things more readable, but for now this will work.

It would be nice if the statsPtr local variable could be removed and
instead index the statsEquip array by indexing with i, but I couldn't
get that to match.