DRA DrawConsumableCount matching (#332)
When the player is viewing the Equip menu, if they have a consumable
item equipped, the quantity that they have is displayed in parentheses
to the right of that item. Here is a random screenshot I found on Google
showing an example of this:
http://www.execulink.com/~maggi/wbw/airback-4.gif

When rendering the Equip menu, this is called twice, once for each equip
hand. It builds a string which has `(`, then the number, and `)`, and
calls the function to draw that string on the menu.

This is my first time working with strings in a function, so let me know
if I've done anything in a non-preferred way.

I believe this is a reasonable name for this function, but it can of
course be changed if that is preferred. However, given that it is only
called in one place and its behavior is clear, I think we know enough
about it to solidly name it this.

I've placed more comments in this code than usual because tracking the
way it builds a string one character at a time is a bit confusing;
hopefully the commenting isn't considered excessive.