DRA CheckChainLimit matching (#336)
When throwable items are used, there is a limit to how many you can
"chain" together. For example, if you equip a Buffalo Star, and press
the throw button as fast as possible, you will throw three of them, and
then you will not be able to throw more until those three despawn.

This function enforces that limit. Given an item ID that could be on
screen, it iterates through g_Entities[16-64] and counts up how many
entities match the one you have thrown. If that number is not below the
chain limit, -1 will be returned, indicating that creating more of them
is not allowed.

A few mysteries still exist. I do not know what the unkBool is, but it
directly corresponds with the top bit of entity->params. I also do not
know what the unkAE is, but it clearly indicates the item ID of a thrown
entity.

Using the previous Equipment struct definition, this code was producing
an `lhu` instruction, where the assembly uses `lbu`. Therefore, I
determine that the `chainable` member should be a u8 instead of a u16. I
have adjusted the struct to match this, and also modified the splat
extraction to show unk17 separately. I will look into what this unk17
might be.

This function is only called by func_8010EDB8. Its purpose seems clear
so I have given it this name.

This is also the final function called by func_8010EDB8 which is not
decompiled.