DRA func_800F72BC matching, rework DRA variables (#337)
**This PR has several things left as "to-do" and should not be merged
as-is. It does match, but some variables should be renamed and functions
should be reworked**

I started off with decompiling func_800F72BC. The scratch for that is
here: https://decomp.me/scratch/z2LZu

Very notably, it loads D_80137692, and then references a value -0x3A
bytes offset from that. This indicates a compiler optimization loading a
second value from a shared struct that contains both D_80137692 and
0x80137658 (which is the second menu in the array of MenuContext
objects). It turns out that creating this struct also allowed the array
to be an actual array of MenuContext items, rather than the hack that
previously existed with "can't use "extern MenuContext D_8013761C[] as
it's 2-byte aligned". We now can have this as an array, because it is
contained in a struct. This should be a useful tip to keep in mind for
the future.

I have packaged all variables from that set of MenuContexts to
D_80137692 into a new struct, which only one exists of. This somewhat
reminds me of g_Status being the only PlayerStatus that exists. For now,
just for functionality sake, the struct is UnkDRAMenuStruct and the only
instance is draMenuStuff - these names are intentionally bad to make
sure we choose proper names for them. I think MenuStatus and
g_MenuStatus could be an acceptable option. I've gone through every
reference to every variable within this range of addresses, and changed
them to reference this struct instead - it all matches.

Back to the function itself: This function renders a portion of the
weapon selection screen. @sozud posted a very nice screenshot of this:
https://media.discordapp.net/attachments/1079395108501331990/1127381120103747744/image.png?width=705&height=499
This renders the readouts on the right side of the screen, including the
words "STR CON INT LCK, all the numbers and arrows to the right of
those, and the arrows and right numbers for ATT and DEF (however, the
numbers left of the arrows on ATT and DEF are handled separately).

This function calls func_800F7218 which evaluates whether the two weapon
options (the one currently equipped, and the one you are hovering over)
are greater than, less than, or equal in the relevant stats. The
function could probably be named statChangeArrow, or something like
that. That function is only ever called by this one, so we should pick a
name now, since it's unlikely anyone will touch it in the future.

This function (the one I have decompiled here) could be called
DrawStatChanges.

This function, confusingly, references DRA variables
g_StatusAttackRightHand, g_StatusAttackLeftHand, and
g_StatusDefenseEquip, and g_StatusPlayerStatsTotal. These are NOT
elements of g_Status, but are copied from elements of g_Status in a
different function. Some light analysis of the statChangeArrow function
indicates that it shows an "increase" arrow when the second argument is
higher. Therefore, it appears that the second argument is the stat that
the player would get if the select the item which the cursor is on.
These variables should therefore probably be renamed to
"proposedAttackRightHand", "newAttackRightHand" or
"cursorAttackRightHand", or something like that. This would be more
descriptive and less confusing.

I think that's everything. There's a lot to go through here, but
hopefully these changes will help our understanding of how the game's
menuing system works.

Oh, and one final thing: The new struct starts with the three menus at
0x8013761C. There are two variables referenced in other places which are
D_80137614 and D_80137618. When I moved the start of the struct to
include these, and referred to them as draMenuStuff.D_80137614, it made
the functions that reference those variables fail to match. That means
the struct should probably start with the menu objects at the beginning
(as I have it here). It is possible that the end of the struct is beyond
where I have it now, but that will require more work to determine. All I
know for sure is that it contains these two members which were both
referenced in this function.