DRA BlinkMenuCursor matching (#343)
This is another one of the menu functions.

I believe it relates to the cursor-blinking animation in the menus,
given that it loads a blink_value from the g_blinkTimer, and stores its
values into a primitive. I've chosen to call it BlinkMenuCursor. There's
a chance I'm wrong, but at least for now this name seems reasonable.

Thank you @Xeeynamo for the final adjustment on args to get this
matching!

I also included a small adjustment to the comments on the MenuData
struct - it's handy to know where all three menus are. This didn't seem
worth having a whole dedicated PR for such a tiny change to only a few
characters, but let me know if you would like it removed.