Decompile DRA func_8011B190 to EntityUnarmedAttack (#348)
This was a tricky one.

This function runs on every frame while the player is performing an
unarmed punch. It does NOT run if the player equips a punching item
(such as the Jewel Knuckles). It is possible that it can run with other
items, but I have not been able to get it to trigger with anything
equipped besides bare hands. I chose to name it EntityUnarmedAttack, but
with a comment, noting that there could be more to this function.

This function required me to create a new struct. As far as I can tell,
no struct like this is used anywhere else in any decompiled function.
Its members control animations and sound. Particularly, it plays a sound
(stored in the struct) when a particular condition (also stored in the
struct) with the player's animations is met. I therefore call it
animSoundEvent for now.

At the end of this function, `func_8010DB38` is called. This appears to
be the only place in the codebase where that function is called. That
function has several lines which resemble the UpdateAnim function that I
decompiled recently, and which exists in TT_000. I have decompiled this
function as well, and will be submitting a separate PR for it in a
moment. I will be renaming that function to UpdateUnarmedAnim, and
therefore the call in this function will need to be refactored. It may
make sense to merge that other function prior to merging this one.