Decompile DRA func_8010DB38 to UpdateUnarmedAnim (#349)
This function is only called in one place. That place is in the function
I have decompiled in a separate PR, EntityUnarmedAttack. This function
appears to update the animation for that entity, and therefore I have
given it this name.

Discovering that this function, which does a subset of what UpdateAnim
does, exists might mean we want to think about looking more deeply into
UpdateAnim, and determining if it should have a more particular name.