Split unk10 to two bytes in Equipment (#353)
I am currently working on func_8010EDB8 (scratch here:
https://decomp.me/scratch/G7vpg) which handles a lot of the processing
that happens when the player uses a weapon.

This appears to be referencing an unk11 member of the Equipment struct,
but previously unk10 was a u16. As far as I can tell, we have not done
anything to reference unk10, and therefore it is reasonable to assume
these are two separate variables.

More work needs to be done, but for now I think the fact that this
function uses unk11 is enough evidence to conclude that they are two
separate values.

I have modified the struct in game.h, as well as modified equipment.py,
to reflect this different organization of values in the struct.