Equipment struct update (#358)
I'm currently working on decompiling func_800FA7E8. Scratch here:
https://decomp.me/scratch/pSeXX

This references unk1C and unk20 in Equipment, and it reads them with the
`lw` instruction, indicating to me that these are each 32-bit values.
unk1E and unk22 do not exist. I'm therefore updating the struct to show
these.

Some preliminary analysis indicates that these variables are related to
the action of pressing both attack buttons at the same time. All items
have 0 for both of these, except that for unk20, Heaven Sword has 1, and
Shield Rod and Mablung Sword have 16382. For unk1C, we have the Heaven
Sword and all the shields. These again correspond to pairing with other
weapons. I will hopefully get a decomp working, but for now, there seems
to be sufficient evidence to change the struct members to match this
configuration.