ST0 func_801A8E60 matching (#359)
This function executes some kind of cutscene script happening together
with the dialogue. There is a bunch of logic that is not actually used
by ST0 and it might have been added to run more complex scenes in other
stages.

@sonicdcer detected `D_801C257C` is used by `EntityDracula` based on
what I read in `st0.h`. But I also found it is used in `EntityCutscene`.
Given this function I decompiled it seems to be some kind of global set
of flags that set or unset when the dialogue wants to trigger specific
actions from the subscribed entities.

The same function is also used in NO3 (Alucard and Death cutscene) and
NZ0 (Alucard and Maria after you pass the blue magic sealed door). I
think it is still a bit too early to decompile the other duplicate
functions as I want to understand more how much the `Dialogue` structure
extends. I still do not know if `D_801C257C` is part of it and there is
a quite big gap of empty data from `D_801C2510`. Based on the future
discoveries I might also move `struct Dialogue` into `stage.h`.