DRA HandlePlay matching (#367)
Very happy of this one. I tried decompiling it two months ago without
success but I decided to give it another try less than 24 hours ago. It
is the function called on every frame while the normal gameplay is
running. This also manages how to transition from a stage to another,
which files to load, set specific random seeds when playing a Tactics
recording.

Note that I refactored `func_800EAF28` as it is used by `HandlePlay`. I
also moved `GameSteps` out of `gameover.c` as it is started to be used
in another handler too. I do not know what `SET_UNK_0B` or `SET_UNK_12`
do, but I added in case at some point we want to explore their meaning.

`func_800F298C` is the function that handles the whole gameplay. That is
the biggest DRA function and the second biggest function of the game. I
hope one day we will manage to decompile that too.

In `case Play_LoadStageChr` there seems to be a hidden loading screen
populated by that `for` loop right after `D_8013640C =
AllocPrimitives(PRIM_GT4, 16);`. It is hidden because `blendMode` is set
to `0`. Setting it to either `8` or `0x88` shows a black screen rather
than showing Alucard frozen in the CD room for those two seconds. I do
not believe it was supposed to be black as X2, Y2, X3 and Y3 have some
kind of rotation and the U/V coords hints the fact some textures were
supposed to be loaded.